feat(wip): bind vars in test scope
This change works towards making vars bound in an auto_assert available
in scope for the rest of the test. The primary challenge is that the
vars to be bound can change when an assertion updates, so you end up
having to bind an old var to a temp value like `:__mneme_no_bound_var__`
or something. Then future assertions that use that var try to update to
use that temp value and things get hairy.

There may be a way to handle this elegantly, but I can't think of one
right now.