fix(diff): disable minimization as it had unintended issues
The minimization optimization intends to reduce the size of the syntax
trees being diffed in a pre-processing step in a way that does not
affect the diff. The minimization strategy being used was incorrect and
resulted in bad diffs in a number of cases.

Minimization is being entirely bypassed at the moment, which results in
a significant slow down in diffing (1.7x-4x depending on size).