fix: error when test file changes before assertion runs
If a test file changes before Mneme gets to it (for example, while it's
waiting on input from an auto-assertion in another file), it won't be
able to find the assertion node to update. This would previously raise
an exception (`Sourceror.Zipper.node/1`), but now will skip the test and
add to a list of tests not saved to be shown at the end of the run.

Closes #53