fix: ensure mneme can find node when updates change line numbers
When Rewrite was refactored in 0.7.0, the behavior of updating a source
subtly changed. Previously, if you updated a source with a non-formatted
quoted form, it would format that form and set the source content to the
formatted string. Now, it will additionally re-parse that newly
formatted content and set the `:quoted` key to the re-parsed source. If
the formatting changed anything, some nodes will differ in
line/column/etc.

Mneme relies on having the original ast with original line numbers to
find assertion nodes in the ast, so this was causing spurious errors
when an update added/removed lines.